@@ -20,7 +20,7 @@ var certificatesCmd = &cobra.Command{
 	Args:    cobra.ExactArgs(1),
 	Run: func(cmd *cobra.Command, args []string) {
 		// Cadence is not a factor in certs, so we hard-code allowUnsuppd to false for the last argument.
-		ar := assetreqs.New(token, "certificates", args[0], "", "",
+		ar := assetreqs.New(token, "certificates", args[0], "", "", "",
 			assetFilePath, assetFileName, outFormat, false)
 		err := ar.GetAsset()
 		if err != nil {
